requirements.txt
requests==2.26.0
discord.py==1.7.3
python-dotenv==0.19.0
Flask==2.0.1
waitress==2.0.0